feat(graphql): add code generation workflow and helper functions for GraphQL session conversion
